Remove security operation apis and combine tests
