+++ src/ResourceManagement/SqlManagement/Sql.Tests/BlobAuditingTest.cs
@@ -1,0 +1,178 @@
+﻿using Microsoft.Azure.Management.Resources;
+using Microsoft.Azure.Management.Sql;
+using Microsoft.Azure.Management.Sql.Models;
+using Microsoft.Azure.Test.HttpRecorder;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace Sql.Tests
+{
+    /// <summary>
+    /// Contains tests for the lifecycle of a server blob auditing policy
+    /// </summary>
+    public class BlobAuditingTest
+    {
+        [Fact]
+        public void TestBlobAuditingApis()
+        {
+            string testPrefix = "server-blob-auditing-test-";
+            string testName = this.GetType().FullName;
+
+            SqlManagementTestUtilities.RunTestInNewV12Server(testName, "TestBlobAuditing", testPrefix,
+                 (resClient, sqlClient, resourceGroup, server) =>
+                {
+                    // create some databases in server
+                    Database[] databases = SqlManagementTestUtilities.CreateDatabasesAsync(
+                       sqlClient, resourceGroup.Name, server, testPrefix, 2).Result;
+
+                    IList<string> auditActionsAndGroups = new List<string> { "SCHEMA_OBJECT_ACCESS_GROUP", "UPDATE on database::testdb by public" };
+
+                    // ******* Server blob auditing *******
+                    // Commented out due to issues with async operation response
+                    //ServerBlobAuditingPolicy defaultServerPolicyResponse = sqlClient.ServerBlobAuditingPolicies.Get(resourceGroup.Name, server.Name);
+
+                    //// Verify that the initial Get request contains the default policy.
+                    //VerifyServerAuditingPolicyInformation(GetDefaultServerBlobAuditingProperties(), defaultServerPolicyResponse);
+
+                    //// Modify the policy properties, send and receive and see it its still ok
+                    //IList<string> auditActionsAndGroups = new List<string> { "SCHEMA_OBJECT_ACCESS_GROUP", "UPDATE on database::testdb by public" };
+                    //ServerBlobAuditingPolicy updatedServerPolicy = new ServerBlobAuditingPolicy
+                    //{
+                    //    State = BlobAuditingPolicyState.Disabled,
+                    //    RetentionDays = 8,
+                    //    StorageAccountAccessKey = "sdlfkjabc+sdlfkjsdlkfsjdfLDKFTERLKFDFKLjsdfksjdflsdkfD2342309432849328476458/3RSD==",
+                    //    StorageEndpoint = "https://MyAccount.blob.core.windows.net/",
+                    //    AuditActionsAndGroups = auditActionsAndGroups,
+                    //    StorageAccountSubscriptionId = "00000000-1234-0000-5678-000000000000",
+                    //    IsStorageSecondaryKeyInUse = false
+                    //};
+
+                    ////Set blob auditing policy for server
+                    //sqlClient.ServerBlobAuditingPolicies.CreateOrUpdate(resourceGroup.Name, server.Name, updatedServerPolicy);
+
+                    ////Get blob auditing server policy
+                    //var getUpdatedServerPolicyResponse = sqlClient.ServerBlobAuditingPolicies.Get(resourceGroup.Name, server.Name);
+
+                    //// Verify that the Get request contains the updated policy.
+                    //VerifyServerAuditingPolicyInformation(updatedServerPolicy, getUpdatedServerPolicyResponse);
+
+                    // ******* Database blob auditing *******
+
+                    string dbName = databases[0].Name;
+                    DatabaseBlobAuditingPolicy defaultDatabasePolicyResponse = sqlClient.DatabaseBlobAuditingPolicies.Get(resourceGroup.Name, server.Name, dbName);
+
+                    // Verify that the initial Get request contains the default policy.
+                    VerifyDatabaseAuditingPolicyInformation(GetDefaultDatabaseBlobAuditingProperties(), defaultDatabasePolicyResponse);
+
+                    // Modify the policy properties, send and receive and see it its still ok
+                    DatabaseBlobAuditingPolicy updatedDatabasePolicy = new DatabaseBlobAuditingPolicy
+                    {
+                        State = BlobAuditingPolicyState.Disabled,
+                        RetentionDays = 5,
+                        StorageAccountAccessKey = "sdlfkjabc+sdlfkjsdlkfsjdfLDKFTERLKFDFKLjsdfksjdflsdkfD2342309432849328476458/3RSD==",
+                        StorageEndpoint = "https://MyAccount.blob.core.windows.net/",
+                        AuditActionsAndGroups = auditActionsAndGroups,
+                        StorageAccountSubscriptionId = "00000000-1234-0000-5678-000000000000",
+                        IsStorageSecondaryKeyInUse = false
+                    };
+                    sqlClient.DatabaseBlobAuditingPolicies.CreateOrUpdate(resourceGroup.Name, server.Name, dbName, updatedDatabasePolicy);
+
+                    var getUpdatedDatabasePolicyResponse = sqlClient.DatabaseBlobAuditingPolicies.Get(resourceGroup.Name, server.Name, dbName);
+                    // Verify that the Get request contains the updated policy.
+                    VerifyDatabaseAuditingPolicyInformation(updatedDatabasePolicy, getUpdatedDatabasePolicyResponse);
+                });
+        }
+
+
+        /// <summary>
+        /// Verify that the received properties match their expected values
+        /// </summary>
+        /// <param name="expected">The expected value of the properties object</param>
+        /// <param name="actual">The properties object that needs to be checked</param>
+        private void VerifyServerAuditingPolicyInformation(ServerBlobAuditingPolicy expected, ServerBlobAuditingPolicy actual)
+        {
+            Assert.Equal(expected.State, actual.State);
+            Assert.Equal(expected.RetentionDays, actual.RetentionDays);
+            Assert.Equal(expected.StorageEndpoint, actual.StorageEndpoint);
+            Assert.Equal(string.Empty, actual.StorageAccountAccessKey);
+            if (expected.AuditActionsAndGroups == null)
+            {
+                Assert.Equal(null, actual.AuditActionsAndGroups);
+            }
+            else
+            {
+                Assert.Equal(expected.AuditActionsAndGroups.Count, actual.AuditActionsAndGroups.Count);
+                actual.AuditActionsAndGroups.ForEach(s => Assert.True(expected.AuditActionsAndGroups.Any(es => es.Equals(s))));
+            }
+            Assert.Equal(expected.StorageAccountSubscriptionId, actual.StorageAccountSubscriptionId);
+            Assert.Equal(expected.IsStorageSecondaryKeyInUse, actual.IsStorageSecondaryKeyInUse);
+        }
+
+        /// <summary>
+        /// Returns a ServerBlobAuditingPolicy object that holds the default settings for a server blob auditing policy
+        /// </summary>
+        /// <returns>A ServerBlobAuditingPolicy object with the default server audit policy settings</returns>
+        private ServerBlobAuditingPolicy GetDefaultServerBlobAuditingProperties()
+        {
+            ServerBlobAuditingPolicy properties = new ServerBlobAuditingPolicy
+            {
+                State = BlobAuditingPolicyState.Disabled,
+                RetentionDays = 0,
+                StorageAccountAccessKey = string.Empty,
+                StorageEndpoint = string.Empty,
+                AuditActionsAndGroups = new List<string>(),
+                StorageAccountSubscriptionId = "00000000-0000-0000-0000-000000000000",
+                IsStorageSecondaryKeyInUse = false,
+            };
+
+            return properties;
+        }
+
+        /// <summary>
+        /// Verify that the received properties match their expected values
+        /// </summary>
+        /// <param name="expected">The expected value of the properties object</param>
+        /// <param name="actual">The properties object that needs to be checked</param>
+        private void VerifyDatabaseAuditingPolicyInformation(DatabaseBlobAuditingPolicy expected, DatabaseBlobAuditingPolicy actual)
+        {
+            Assert.Equal(expected.State, actual.State);
+            Assert.Equal(expected.RetentionDays, actual.RetentionDays);
+            Assert.Equal(expected.StorageEndpoint, actual.StorageEndpoint);
+            Assert.Equal(string.Empty, actual.StorageAccountAccessKey);
+            if (expected.AuditActionsAndGroups == null)
+            {
+                Assert.Equal(null, actual.AuditActionsAndGroups);
+            }
+            else
+            {
+                Assert.Equal(expected.AuditActionsAndGroups.Count, actual.AuditActionsAndGroups.Count);
+                actual.AuditActionsAndGroups.ForEach(s => Assert.True(expected.AuditActionsAndGroups.Any(es => es.Equals(s))));
+            }
+            Assert.Equal(expected.StorageAccountSubscriptionId, actual.StorageAccountSubscriptionId);
+            Assert.Equal(expected.IsStorageSecondaryKeyInUse, actual.IsStorageSecondaryKeyInUse);
+        }
+
+        /// <summary>
+        /// Returns a BlobAuditingProperties object that holds the default settings for a database blob auditing policy
+        /// </summary>
+        /// <returns>A BlobAuditingProperties object with the default database audit policy settings</returns>
+        private DatabaseBlobAuditingPolicy GetDefaultDatabaseBlobAuditingProperties()
+        {
+            DatabaseBlobAuditingPolicy properties = new DatabaseBlobAuditingPolicy
+            {
+                State = BlobAuditingPolicyState.Disabled,
+                RetentionDays = 0,
+                StorageAccountAccessKey = string.Empty,
+                StorageEndpoint = string.Empty,
+                AuditActionsAndGroups = new List<string>(),
+                StorageAccountSubscriptionId = "00000000-0000-0000-0000-000000000000",
+                IsStorageSecondaryKeyInUse = false,
+            };
+
+            return properties;
+        }
+    }
+}
