--- src/ResourceManagement/SqlManagement/Sql.Tests/ServerBlobAuditingTest.cs
@@ -1,132 +1,0 @@
-﻿using Microsoft.Azure.Management.Resources;
-using Microsoft.Azure.Management.Sql;
-using Microsoft.Azure.Management.Sql.Models;
-using Microsoft.Azure.Test.HttpRecorder;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Xunit;
-
-namespace Sql.Tests
-{
-    /// <summary>
-    /// Contains tests for the lifecycle of a server blob auditing policy
-    /// </summary>
-    public class ServerBlobAuditingTest
-    {
-        [Fact]
-        public void TestServerBlobAuditingApis()
-        {
-            string testPrefix = "server-blob-auditing-test-";
-            string testName = this.GetType().FullName;
-            const int retriesCount = 5;
-            const int databasesCount = 12;
-
-            SqlManagementTestUtilities.RunTestInNewV12Server(testName, "TestServerBlobAuditing", testPrefix,
-                 (resClient, sqlClient, resourceGroup, server) =>
-                {
-                    //create server with a lot of databases
-                    for (int i = 0; i < databasesCount; i++)
-                    {
-                        string dbName = string.Format("db{0}", i);
-                        sqlClient.Databases.CreateOrUpdate(resourceGroup.Name, server.Name, dbName, new Database()
-                        {
-                            Location = server.Location,
-                        });
-                    }
-
-                    ServerBlobAuditingPolicy defaultServerPolicyResponse = sqlClient.ServerBlobAuditingPolicies.Get(resourceGroup.Name, server.Name);
-
-                    // Verify that the initial Get request contains the default policy.
-                    VerifyAuditingPolicyInformation(GetDefaultBlobAuditgProperties(), defaultServerPolicyResponse);
-
-                    // Modify the policy properties, send and receive and see it its still ok
-                    IList<string> auditActionsAndGroups = new List<string> { "SCHEMA_OBJECT_ACCESS_GROUP", "UPDATE on database::testdb by public" };
-                    ServerBlobAuditingPolicy updatedServerPolicy = new ServerBlobAuditingPolicy
-                    {
-                        State = SecurityConstants.c_DisabledValue,
-                        RetentionDays = 8,
-                        StorageAccountAccessKey = "sdlfkjabc+sdlfkjsdlkfsjdfLDKFTERLKFDFKLjsdfksjdflsdkfD2342309432849328476458/3RSD==",
-                        StorageEndpoint = "https://MyAccount.blob.core.windows.net/",
-                        AuditActionsAndGroups = auditActionsAndGroups,
-                        StorageAccountSubscriptionId = "00000000-1234-0000-5678-000000000000",
-                        IsStorageSecondaryKeyInUse = false
-                    };
-                    
-                    //Set blob auditing policy for server
-                    sqlClient.ServerBlobAuditingPolicies.CreateOrUpdate(resourceGroup.Name, server.Name, updatedServerPolicy);
-                    //Waiting for the operation to start
-                    Task.Delay(TimeSpan.FromSeconds(2));
-
-                    BlobAuditingOperationListResult operationResults = sqlClient.ServerBlobAuditingPolicies.ListOperationResults(resourceGroup.Name, server.Name);
-                    if (operationResults != null && operationResults.Value.Any())
-                    {
-                        //The set policy operation has not finished yet
-                        BlobAuditingOperationResult operationStatusResponse = operationResults.Value.First();
-                        string operationId = operationStatusResponse.OperationId;
-                        for (int i = 0; i < retriesCount; i++)
-                        {
-                            operationStatusResponse = sqlClient.ServerBlobAuditingPolicies.GetOperationResult(
-                                    resourceGroup.Name, server.Name, operationId);
-                            if (operationStatusResponse != null && operationStatusResponse.State == SecurityConstants.c_SucceededOperationStatus)
-                            {
-                                break;
-                            }
-                            Task.Delay(TimeSpan.FromSeconds(5));
-                        }
-                    }
-
-                    //Get blob auditing server policy
-                    var getUpdatedPolicyResponse = sqlClient.ServerBlobAuditingPolicies.Get(resourceGroup.Name, server.Name);
-                    // Verify that the Get request contains the updated policy.
-                    VerifyAuditingPolicyInformation(updatedServerPolicy, getUpdatedPolicyResponse);
-                });
-        }
-
-
-        /// <summary>
-        /// Verify that the received properties match their expected values
-        /// </summary>
-        /// <param name="expected">The expected value of the properties object</param>
-        /// <param name="actual">The properties object that needs to be checked</param>
-        private void VerifyAuditingPolicyInformation(ServerBlobAuditingPolicy expected, ServerBlobAuditingPolicy actual)
-        {
-            Assert.Equal(expected.State, actual.State);
-            Assert.Equal(expected.RetentionDays, actual.RetentionDays);
-            Assert.Equal(expected.StorageEndpoint, actual.StorageEndpoint);
-            Assert.Equal(string.Empty, actual.StorageAccountAccessKey);
-            if (expected.AuditActionsAndGroups == null)
-            {
-                Assert.Equal(null, actual.AuditActionsAndGroups);
-            }
-            else
-            {
-                Assert.Equal(expected.AuditActionsAndGroups.Count, actual.AuditActionsAndGroups.Count);
-                actual.AuditActionsAndGroups.ForEach(s => Assert.True(expected.AuditActionsAndGroups.Any(es => es.Equals(s))));
-            }
-            Assert.Equal(expected.StorageAccountSubscriptionId, actual.StorageAccountSubscriptionId);
-            Assert.Equal(expected.IsStorageSecondaryKeyInUse, actual.IsStorageSecondaryKeyInUse);
-        }
-
-        /// <summary>
-        /// Returns a ServerBlobAuditingPolicy object that holds the default settings for a server blob auditing policy
-        /// </summary>
-        /// <returns>A ServerBlobAuditingPolicy object with the default server audit policy settings</returns>
-        private ServerBlobAuditingPolicy GetDefaultBlobAuditgProperties()
-        {
-            ServerBlobAuditingPolicy properties = new ServerBlobAuditingPolicy
-            {
-                State = SecurityConstants.c_DisabledValue,
-                RetentionDays = 0,
-                StorageAccountAccessKey = string.Empty,
-                StorageEndpoint = string.Empty,
-                AuditActionsAndGroups = new List<string>(),
-                StorageAccountSubscriptionId = "00000000-0000-0000-0000-000000000000",
-                IsStorageSecondaryKeyInUse = false,
-            };
-
-            return properties;
-        }
-    }
-}
