--- src/ResourceManagement/SqlManagement/Sql.Tests/Utilities/SecurityConstants.cs
@@ -1,8 +1,0 @@
-﻿namespace Sql.Tests
-{
-    public class SecurityConstants
-    {
-        public const string c_DisabledValue = "Disabled";
-        public const string c_SucceededOperationStatus = "1";
-    }
-}
